+++ chainmail/src/main/java/ninjaphenix/chainmail/impl/mixinhelpers/ItemGroupArrayExpander.java
@@ -1,0 +1,6 @@
+package ninjaphenix.chainmail.impl.mixinhelpers;
+
+public interface ItemGroupArrayExpander
+{
+    int chainmail_expandArraySize();
+}