+++ creativebuttonmover/src/main/java/ninjaphenix/creativebuttonmover/Vec2i.java
@@ -1,0 +1,13 @@
+package ninjaphenix.creativebuttonmover;
+
+public class Vec2i
+{
+    public final int x;
+    public final int y;
+
+    public Vec2i(final int x, final int y)
+    {
+        this.x = x;
+        this.y = y;
+    }
+}