+++ expandedstorage/src/main/java/ninjaphenix/expandedstorage/content/ModBlocks.java
@@ -10,6 +10,8 @@ import net.minecraft.util.Identifier;
 import net.minecraft.util.registry.Registry;
 import ninjaphenix.expandedstorage.ExpandedStorage;
 import ninjaphenix.expandedstorage.api.Registries;
+import ninjaphenix.expandedstorage.api.Registries.ChestTierData;
+import ninjaphenix.expandedstorage.api.Registries.TierData;
 import ninjaphenix.expandedstorage.block.CursedChestBlock;
 import ninjaphenix.expandedstorage.block.OldChestBlock;
 
@@ -45,7 +47,7 @@ public final class ModBlocks
         final Identifier id = ExpandedStorage.getId("old_" + name);
         Registry.register(Registry.BLOCK, id, block);
         Registry.register(Registry.ITEM, id, new BlockItem(block, new Item.Settings().group(ExpandedStorage.group)));
-        Registries.OLD_CHEST.add(ExpandedStorage.getId(name), new Registries.TierData(rows * 9, containerName, id));
+        Registry.register(Registries.OLD_CHEST, ExpandedStorage.getId(name), new TierData(rows * 9, containerName, id));
         return block;
     }
 
@@ -53,14 +55,11 @@ public final class ModBlocks
     {
         final CursedChestBlock block = new CursedChestBlock(Block.Settings.copy(material));
         final Text containerName = new TranslatableText("container.expandedstorage." + name);
-        final Identifier singleTexture = ExpandedStorage.getId("entity/" + name + "/single");
-        final Identifier vanillaTexture = ExpandedStorage.getId("entity/" + name + "/vanilla");
-        final Identifier tallTexture = ExpandedStorage.getId("entity/" + name + "/tall");
-        final Identifier longTexture = ExpandedStorage.getId("entity/" + name + "/long");
         final Identifier id = ExpandedStorage.getId(name);
         Registry.register(Registry.BLOCK, id, block);
         Registry.register(Registry.ITEM, id, new BlockItem(block, new Item.Settings().group(ExpandedStorage.group)));
-        Registries.CHEST.add(id, new Registries.ChestTierData(rows * 9, containerName, id, singleTexture, vanillaTexture, tallTexture, longTexture));
+        Registry.register(Registries.CHEST, id, new ChestTierData(rows * 9, containerName, id,
+                type -> ExpandedStorage.getId(String.format("entity/%s/%s", name, type.asString()))));
         return block;
     }