+++ expandedstorage/src/main/java/ninjaphenix/expandedstorage/item/ChestConversionItem.java
@@ -105,8 +105,6 @@ public class ChestConversionItem extends ChestModifierItem
     @SuppressWarnings("ConstantConditions")
     protected ActionResult useModifierOnBlock(ItemUsageContext context, BlockState state)
     {
-        // todo fix this for other mods.
-        // Pehaps allow mods to define equivalents or use tags somehow e.g. Tag<Identifier>("expandedstorage:wood")
         if (state.getBlock() == Blocks.CHEST && from.equals(ExpandedStorage.getId("wood_chest")))
         {
             World world = context.getWorld();