+++ ninjaphenix-container-library/src/main/java/ninjaphenix/containerlib/impl/client/screen/SingleScreen.java
@@ -1,6 +1,7 @@
 package ninjaphenix.containerlib.impl.client.screen;
 
 
+import net.fabricmc.loader.api.FabricLoader;
 import ninjaphenix.containerlib.api.screen.SingleScreenMeta;
 import ninjaphenix.containerlib.api.client.screen.AbstractScreen;
 import ninjaphenix.containerlib.api.client.screen.widget.ScreenTypeSelectionScreenButton;
@@ -9,6 +10,7 @@ import ninjaphenix.containerlib.impl.inventory.SingleContainer;
 public class SingleScreen<T extends SingleContainer> extends AbstractScreen<T, SingleScreenMeta>
 {
     private Rectangle blankArea = null;
+    private ScreenTypeSelectionScreenButton screenSelectButton;
 
     public SingleScreen(T container)
     {
@@ -21,7 +23,9 @@ public class SingleScreen<T extends SingleContainer> extends AbstractScreen<T, S
     protected void init()
     {
         super.init();
-        addButton(new ScreenTypeSelectionScreenButton(x + containerWidth - 19, y + 4));
+        int settingsXOffset = -19;
+        if (FabricLoader.getInstance().isModLoaded("inventorysorter")) { settingsXOffset -= 18; }
+        screenSelectButton = addButton(new ScreenTypeSelectionScreenButton(x + containerWidth + settingsXOffset, y + 4));
         final int blanked = SCREEN_META.BLANK_SLOTS;
         if (blanked > 0)
         {
@@ -29,6 +33,13 @@ public class SingleScreen<T extends SingleContainer> extends AbstractScreen<T, S
             blankArea = new Rectangle(x + xOffset, y + containerHeight - 115, blanked * 18, 18,
                     xOffset, containerHeight, SCREEN_META.TEXTURE_WIDTH, SCREEN_META.TEXTURE_HEIGHT);
         }
+    }
+
+    @Override
+    public void render(int mouseX, int mouseY, float delta)
+    {
+        super.render(mouseX, mouseY, delta);
+        screenSelectButton.renderTooltip(mouseX, mouseY, this::renderTooltip);
     }
 
     @Override