+++ noncorrelatedextras/src/main/java/ninjaphenix/noncorrelatedextras/features/StructureCompassFeature.java
@@ -2,6 +2,7 @@ package ninjaphenix.noncorrelatedextras.features;
 
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemGroup;
+import net.minecraft.util.Identifier;
 import net.minecraft.util.registry.Registry;
 import ninjaphenix.noncorrelatedextras.Main;
 import ninjaphenix.noncorrelatedextras.config.Config;
@@ -9,7 +10,6 @@ import ninjaphenix.noncorrelatedextras.core.Feature;
 import ninjaphenix.noncorrelatedextras.core.ItemAdder;
 import ninjaphenix.noncorrelatedextras.items.StructureCompassItem;
 
-import java.util.Locale;
 import java.util.Set;
 
 public class StructureCompassFeature extends Feature implements ItemAdder
@@ -17,12 +17,11 @@ public class StructureCompassFeature extends Feature implements ItemAdder
     @Override
     public void registerItems()
     {
-        final Set<String> structures = Config.INSTANCE.getEnabledStructureCompasses();
+        final Set<Identifier> structures = Config.INSTANCE.getEnabledStructureCompasses();
         final Item.Settings settings = new Item.Settings().maxCount(1).group(ItemGroup.TOOLS);
-        for (String structure : structures)
+        for (Identifier structure : structures)
         {
-            structure = structure.toLowerCase(Locale.ROOT);
-            StructureCompassItem item = new StructureCompassItem(settings, structure);
+            StructureCompassItem item = new StructureCompassItem(settings, structure.getPath());
             Registry.register(Registry.ITEM, Main.getId(structure + "_compass"), item);
         }
     }